[DO NOT MERGE] Mock verified CIM digest for layers
Verified CIMs will allow the gcs-sidecar to query the root digest for each block CIM and
then validate that against the policy to see if that layer is allowed. The layer CIMs will
be merge mounted only if all of the root digests of all layer CIMs are successfully
validated against the policy.
However, verified CIMs aren't available yet. In order to unblock testing of the policy
engine, this commit mocks the root digest of a block CIM by generating a SHA256 of the
layer path on the host. As long as the layer path remains the same (i.e we won't remove
and repull the same image) the layer digest will remain same and we can use that in the
policy.
Note that this only a temporary change and it shouldn't be merged into main. Once verified
CIMs are ready, we won't need to pass a digest in the mount block CIM request, instead
gcs-sidecar will directly query the digest from the CIM.

Signed-off-by: Amit Barve <ambarve@microsoft.com>